user registration form

+++ studybud/base/views.py
@@ -1,10 +1,13 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
-from .models import Room, Topic
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic
 from .forms import RoomForm
+from django.http import HttpResponse
 
 
 # rooms = [
@@ -14,10 +17,14 @@ from .forms import RoomForm
 # ]
 
 def loginpage(request):
-    context = {}
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
-        username = request.POST.get("username")
+        username = request.POST.get("username").lower()
         password = request.POST.get("password")
 
         try:
@@ -30,16 +37,37 @@ def loginpage(request):
         if user is not None:
             login(request, user=user)
             return redirect('home')
-
         else:
             messages.error(request, "Username or password does not exist")
 
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 
 def logoutuser(request):
     logout(request) 
     return redirect('home')
+
+def registerpage(request):
+
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occured during registration")
+
+    context = {
+        'form': form
+    }
+
+    return render(request, 'base/login_register.html', context)
 
 
 def home(request):
@@ -72,9 +100,9 @@ def room(request, pk):
 
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='/login')
 def createroom(request):
-   
+
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -88,10 +116,14 @@ def createroom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def updateroom(request, pk):
    
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -104,8 +136,13 @@ def updateroom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def deleteroom(request, pk):
+    
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Your are not allowed here")
     
     if request.method == 'POST':
         room.delete()
